Improved Cursed Wool explosive functionality and retextured and add ores for Filad.

+++ src/main/java/net/stilt/cursedmobsandpvp/event/CursedWoolAttackEventHandler.java
@@ -7,6 +7,7 @@ import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
+import net.minecraft.world.GameMode;
 import net.minecraft.world.World;
 import net.minecraft.world.World.ExplosionSourceType;
 import net.stilt.cursedmobsandpvp.block.ModBlocks;
@@ -16,6 +17,7 @@ public class CursedWoolAttackEventHandler implements AttackBlockCallback{
     @Override
     public ActionResult interact(PlayerEntity player, World world, Hand hand, BlockPos pos, Direction direction) {
         BlockState state = world.getBlockState(pos);
+        GameMode gameMode = GameMode.DEFAULT;
             /* Manual spectator check is necessary because AttackBlockCallbacks
                fire before the spectator check */
             if (!player.isSpectator() && state.getBlock() == ModBlocks.Cursed_Wool) {
@@ -23,7 +25,7 @@ public class CursedWoolAttackEventHandler implements AttackBlockCallback{
                 world.createExplosion(null, pos.getX(), pos.getY() + 1, pos.getZ(), 100.0f, ExplosionSourceType.NONE);
                 world.createExplosion(null, pos.getX(), pos.getY() - 1, pos.getZ(), 100.0f, ExplosionSourceType.NONE);
 
-                if (player.getHealth() != 0f){
+                if (player.getHealth() != 0f && !player.getAbilities().creativeMode){
                     player.kill();
                 }
